Shuffle stuff around until tests work again
