+++ frontend/src/session/utils.tsx
@@ -1,6 +1,5 @@
-import { stringify } from "csv-stringify/sync";
 import { Message as RawMessage, Timestamp } from "../types";
-import { Message, Session } from "./types";
+import { Message } from "./types";
 
 export const processTimestamp = ({ timestamp }: Timestamp) =>
   new Date(timestamp);
@@ -14,34 +13,3 @@ export const processMessage = ({
   timestamp: processTimestamp(timestamp),
   content,
 });
-
-export const generateCsv = (session: Session): string => {
-  const rows = [
-    ["Session name", session.name],
-    ["Created at", session.created.toISOString()],
-  ];
-
-  if (session.closed) {
-    rows.push(["Closed at", session.closed.toISOString()]);
-  }
-
-  rows.push([], ["Feedback type", "Content", "Time"]);
-
-  for (const message of session.messages) {
-    let type;
-    let content;
-
-    if (`text` in message.content) {
-      type = `Text message`;
-      content = message.content.text;
-    } else {
-      type = `Rating`;
-      content = String(message.content.rating);
-    }
-
-    const row = [type, content, message.timestamp.toISOString()];
-    rows.push(row);
-  }
-
-  return stringify(rows);
-};
